git-tfs-id: [http://192.168.56.101:8080/tfs/DefaultCollection]$/TikiEngine 2.0;C713

+++ src/Core/Camera.cpp
@@ -38,7 +38,7 @@ namespace TikiEngine
 
 			this->matrices.ProjectionMatrix = Matrix::Transpose(Matrix::CreatePerspectiveFieldOfView(
 				MATH_PI / 4,				
-				(vp->Width / vp->Width),
+				(float)vp->Width / vp->Width,
 				0.01f,
 				1000.0f
 			));
